requirements.txt
Django>=3.0
psycopg2-binary>=2.8
djangorestframework
Pillow==8.1.0
django-cors-headers
gunicorn
django-heroku
drf-nested-routers